Work toward refactoring the time machine layout into a base class and a subclass.
